(1.0.0.127) http://bugzserver/show_bug.cgi?id=26409
git-svn-id: svn://fileserver/activex/AVS/Sources/TeamlabOffice/trunk/ServerComponents@60147 954022d7-b5bf-4e40-9824-e11837661b57

+++ ASCOfficeDocFile/DocxDocConverter/OfficeArt/BlipFactory.h
@@ -92,7 +92,7 @@ namespace OfficeArt
 
 				//узнаем размер файла, и выделяем память в строке
 				xfile.seekg( 0, ios_base::end );
-				xstr.resize( xfile.tellg() );
+				xstr.resize( (size_t)xfile.tellg() );
 				xfile.seekg( 0, ios_base::beg );
 
 				//копируем данные
@@ -197,7 +197,7 @@ namespace OfficeArt
 
 				//узнаем размер файла, и выделяем память в строке
 				xfile.seekg( 0, ios_base::end );
-				xstr.resize( xfile.tellg() );
+				xstr.resize( (size_t)xfile.tellg() );
 				xfile.seekg( 0, ios_base::beg );
 
 				//копируем данные
